requirements/requirements-dev.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile --output-file=requirements/requirements-dev.txt requirements/requirements-dev.in
#
amqp==5.0.6
    # via kombu
asgiref==3.4.1
    # via django
astroid==2.7.3
    # via pylint
attrs==21.2.0
    # via pytest
beautifulsoup4==4.9.3
    # via -r requirements/requirements.in
billiard==3.6.4.0
    # via celery
celery[redis]==5.1.2
    # via
    #   -r requirements/requirements.in
    #   pytest-celery
click==7.1.2
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
click-didyoumean==0.0.3
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.2.0
    # via celery
coverage[toml]==5.5
    # via -r requirements/requirements-dev.in
cssselect==1.1.0
    # via pyquery
django==3.2.7
    # via
    #   -r requirements/requirements.in
    #   django-appconf
    #   django-debug-toolbar
    #   django-extensions
    #   django-recaptcha
    #   django-redis
    #   django-taggit
django-appconf==1.0.4
    # via django-compressor
django-compressor==2.4.1
    # via -r requirements/requirements.in
django-debug-toolbar==3.2.2
    # via -r requirements/requirements-dev.in
django-environ-2==2.2.0
    # via -r requirements/requirements.in
django-extensions==3.1.3
    # via -r requirements/requirements-dev.in
django-recaptcha==2.0.6
    # via -r requirements/requirements.in
django-redis==5.0.0
    # via -r requirements/requirements.in
django-taggit==1.5.1
    # via -r requirements/requirements.in
factory-boy==3.2.0
    # via -r requirements/requirements-dev.in
faker==8.12.1
    # via
    #   -r requirements/requirements-dev.in
    #   factory-boy
flake8==3.9.2
    # via
    #   -r requirements/requirements-dev.in
    #   flake8-builtins
    #   flake8-docstrings
flake8-blind-except==0.2.0
    # via -r requirements/requirements-dev.in
flake8-builtins==1.5.3
    # via -r requirements/requirements-dev.in
flake8-docstrings==1.6.0
    # via -r requirements/requirements-dev.in
flake8-import-order==0.18.1
    # via -r requirements/requirements-dev.in
gunicorn==20.1.0
    # via -r requirements/requirements.in
iniconfig==1.1.1
    # via pytest
isort==5.9.3
    # via pylint
kombu==5.1.0
    # via celery
lazy-object-proxy==1.6.0
    # via astroid
lxml==4.6.3
    # via pyquery
mccabe==0.6.1
    # via
    #   flake8
    #   pylint
packaging==21.0
    # via pytest
pillow==8.3.1
    # via -r requirements/requirements.in
platformdirs==2.3.0
    # via pylint
pluggy==1.0.0
    # via pytest
prompt-toolkit==3.0.20
    # via click-repl
py==1.10.0
    # via pytest
pycodestyle==2.7.0
    # via
    #   flake8
    #   flake8-import-order
pydocstyle==6.1.1
    # via flake8-docstrings
pyflakes==2.3.1
    # via flake8
pylint==2.10.2
    # via
    #   -r requirements/requirements-dev.in
    #   pylint-django
    #   pylint-plugin-utils
pylint-django==2.4.4
    # via -r requirements/requirements-dev.in
pylint-plugin-utils==0.6
    # via pylint-django
pyparsing==2.4.7
    # via packaging
pyquery==1.4.3
    # via -r requirements/requirements.in
pytest==6.2.5
    # via
    #   -r requirements/requirements-dev.in
    #   pytest-django
pytest-celery==0.0.0
    # via -r requirements/requirements-dev.in
pytest-django==4.4.0
    # via -r requirements/requirements-dev.in
python-dateutil==2.8.2
    # via faker
pytz==2021.1
    # via
    #   celery
    #   django
rcssmin==1.0.6
    # via django-compressor
redis==3.5.3
    # via
    #   celery
    #   django-redis
rjsmin==1.1.0
    # via django-compressor
six==1.16.0
    # via
    #   click-repl
    #   django-compressor
    #   python-dateutil
snowballstemmer==2.1.0
    # via pydocstyle
soupsieve==2.2.1
    # via beautifulsoup4
sqlparse==0.4.1
    # via
    #   django
    #   django-debug-toolbar
text-unidecode==1.3
    # via faker
toml==0.10.2
    # via
    #   coverage
    #   pylint
    #   pytest
vine==5.0.0
    # via
    #   amqp
    #   celery
    #   kombu
wcwidth==0.2.5
    # via prompt-toolkit
wrapt==1.12.1
    # via astroid

# The following packages are considered to be unsafe in a requirements file:
# setuptools
